add: TermsContent as a common component used by both admin and ui terms
